backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-dotenv
openai
pydantic
python-multipart
corsmiddleware
